requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
Pillow==12.0.0
django-filter==25.2
django-environ==0.12.0
django-cors-headers==4.9.0
django-model-utils==5.0.0
djangorestframework==3.16.1
django-extensions==4.1
djangorestframework-simplejwt==5.5.1
django-countries==8.1.1